test: split admin v1 integration tests below 600-line cap
Move DB-backed §S1.6 audit invariant tests into a new file
`integration_admin_v1_audit_tests.rs` (550 lines) and trim
`integration_admin_v1_tests.rs` to fast middleware-only tests (209 lines).
Both files share identical helper functions; no test logic changed.
